Create Lcm
Write a program to find LCM of two numbers is discussed here. Two numbers are obtained as input and the prime factors of both the numbers are found. The product of the union of prime factors of both the numbers gives the LCM of the two numbers.

Input & Output Format:

Input consists of two integers.

Output consists of the lcm of two numbers.

Sample Input:

5

30

Sample Output:

LCM of 5 and 30 is 30